Move all tests to `HttpRequestBuilderTest`
